Add batch-oriented base Producer class
The message-oriented model of AIOKafkaProducer is user-friendly and
requires little knowledge of the library's innards to use effectively.
It also limits performance by delegating all batch controls to the
MessageAccumulator, which is hidden entirely from user-interaction,
and creating futures for every single message sent.

This diff makes the following major changes:
- Leaves the AIOKafkaProducer interface intact. All existing use-cases
  should continue to work without modification.
- Adds AIOKafkaBaseProducer as a batch-oriented parent class with the
  primary interfaces of `create_batch()`, `get_batch()`, and
  `send_batch()`. Users of these interfaces operate on MessageBatch
  objects.
- Eliminates MessageAccumulator entirely. Its logic has been pulled into
  AIOKafkaBaseProducer with closer proximity to `sender_routine()`.

It also makes the following minor changes:
- Moves recognizes API versions into a class variable.
- Sets `self._closed = True` earlier in `producer.stop()` to ensure
  messages can't be added while awaiting other shutdown coroutines.
- Tweaks `MessageBatch.wait_drain()` to return an `asyncio.wait()`
  future like `MessageBatch.wait_deliver()`.
- Adds an `api_version` property mapping to `self.client.api_version`.
- Rewrites several `assert` statements as `if/raise` value checks.
- Pulls all existing tests from `tests/test_message_accumulator.py` into
  `tests/test_producer.py` with only minor updates.
- Eliminates the `_PRODUCER_CLIENT_ID_SEQUENCE` logic. Because the
  operation took place at the instance level, the class variable was
  never overwritten and every client simply ended up a value of `1`.
