+++ aiokafka/producer/__init__.py
@@ -1,0 +1,8 @@
+from .base import AIOKafkaBaseProducer  # noqa
+from .base import MessageBatch  # noqa
+from .simple import AIOKafkaProducer  # noqa
+
+__all__ = [
+    "AIOKafkaBaseProducer",
+    "AIOKafkaProducer",
+]
