+++ aiokafka/producer/simple.py
@@ -1,0 +1,265 @@
+import asyncio
+import logging
+
+from kafka.partitioner.default import DefaultPartitioner
+from kafka.protocol.message import Message, MessageSet
+
+from aiokafka.errors import (
+    KafkaTimeoutError,
+    MessageSizeTooLargeError,
+    ProducerClosed,
+    UnknownPartition,
+)
+from aiokafka.producer import AIOKafkaBaseProducer
+from aiokafka.structs import TopicPartition
+
+log = logging.getLogger(__name__)
+
+
+class AIOKafkaProducer(AIOKafkaBaseProducer):
+    """A Kafka client that publishes records to the Kafka cluster.
+
+    The producer consists of a pool of buffer space that holds records that
+    haven't yet been transmitted to the server as well as a background task
+    that is responsible for turning these records into requests and
+    transmitting them to the cluster.
+
+    The send() method is asynchronous. When called it adds the record to a
+    buffer of pending record sends and immediately returns. This allows the
+    producer to batch together individual records for efficiency.
+
+    The 'acks' config controls the criteria under which requests are considered
+    complete. The "all" setting will result in waiting for all replicas to
+    respond, the slowest but most durable setting.
+
+    The key_serializer and value_serializer instruct how to turn the key and
+    value objects the user provides into bytes.
+
+    Arguments:
+        bootstrap_servers: 'host[:port]' string (or list of 'host[:port]'
+            strings) that the producer should contact to bootstrap initial
+            cluster metadata. This does not have to be the full node list.
+            It just needs to have at least one broker that will respond to a
+            Metadata API Request. Default port is 9092. If no servers are
+            specified, will default to localhost:9092.
+        client_id (str): a name for this client. This string is passed in
+            each request to servers and can be used to identify specific
+            server-side log entries that correspond to this client.
+            Default: 'aiokafka-producer-#' (appended with a unique number
+            per instance)
+        key_serializer (callable): used to convert user-supplied keys to bytes
+            If not None, called as f(key), should return bytes. Default: None.
+        value_serializer (callable): used to convert user-supplied message
+            values to bytes. If not None, called as f(value), should return
+            bytes. Default: None.
+        acks (0, 1, 'all'): The number of acknowledgments the producer requires
+            the leader to have received before considering a request complete.
+            This controls the durability of records that are sent. The
+            following settings are common:
+
+            0: Producer will not wait for any acknowledgment from the server
+                at all. The message will immediately be added to the socket
+                buffer and considered sent. No guarantee can be made that the
+                server has received the record in this case, and the retries
+                configuration will not take effect (as the client won't
+                generally know of any failures). The offset given back for each
+                record will always be set to -1.
+            1: The broker leader will write the record to its local log but
+                will respond without awaiting full acknowledgement from all
+                followers. In this case should the leader fail immediately
+                after acknowledging the record but before the followers have
+                replicated it then the record will be lost.
+            all: The broker leader will wait for the full set of in-sync
+                replicas to acknowledge the record. This guarantees that the
+                record will not be lost as long as at least one in-sync replica
+                remains alive. This is the strongest available guarantee.
+
+            If unset, defaults to acks=1.
+        compression_type (str): The compression type for all data generated by
+            the producer. Valid values are 'gzip', 'snappy', 'lz4', or None.
+            Compression is of full batches of data, so the efficacy of batching
+            will also impact the compression ratio (more batching means better
+            compression). Default: None.
+        max_batch_size (int): Maximum size of buffered data per partition.
+            After this amount `send` coroutine will block until batch is
+            drained.
+            Default: 16384
+        linger_ms (int): The producer groups together any records that arrive
+            in between request transmissions into a single batched request.
+            Normally this occurs only under load when records arrive faster
+            than they can be sent out. However in some circumstances the client
+            may want to reduce the number of requests even under moderate load.
+            This setting accomplishes this by adding a small amount of
+            artificial delay; that is, if first request is processed faster,
+            than `linger_ms`, producer will wait `linger_ms - process_time`.
+            This setting defaults to 0 (i.e. no delay).
+        partitioner (callable): Callable used to determine which partition
+            each message is assigned to. Called (after key serialization):
+            partitioner(key_bytes, all_partitions, available_partitions).
+            The default partitioner implementation hashes each non-None key
+            using the same murmur2 algorithm as the Java client so that
+            messages with the same key are assigned to the same partition.
+            When a key is None, the message is delivered to a random partition
+            (filtered to partitions with available leaders only, if possible).
+        max_request_size (int): The maximum size of a request. This is also
+            effectively a cap on the maximum record size. Note that the server
+            has its own cap on record size which may be different from this.
+            This setting will limit the number of record batches the producer
+            will send in a single request to avoid sending huge requests.
+            Default: 1048576.
+        metadata_max_age_ms (int): The period of time in milliseconds after
+            which we force a refresh of metadata even if we haven't seen any
+            partition leadership changes to proactively discover any new
+            brokers or partitions. Default: 300000
+        request_timeout_ms (int): Produce request timeout in milliseconds.
+            As it's sent as part of ProduceRequest (it's a blocking call),
+            maximum waiting time can be up to 2 * request_timeout_ms.
+            Default: 40000.
+        retry_backoff_ms (int): Milliseconds to backoff when retrying on
+            errors. Default: 100.
+        api_version (str): specify which kafka API version to use.
+            If set to 'auto', will attempt to infer the broker version by
+            probing various APIs. Default: auto
+        security_protocol (str): Protocol used to communicate with brokers.
+            Valid values are: PLAINTEXT, SSL. Default: PLAINTEXT.
+        ssl_context (ssl.SSLContext): pre-configured SSLContext for wrapping
+            socket connections. Directly passed into asyncio's
+            `create_connection`_. For more information see :ref:`ssl_auth`.
+            Default: None.
+
+    Note:
+        Many configuration parameters are taken from the Java client:
+        https://kafka.apache.org/documentation.html#producerconfigs
+    """
+
+    def __init__(self, *args, partitioner=DefaultPartitioner(),
+                 key_serializer=None, value_serializer=None, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._partitioner = partitioner
+        self._key_serializer = key_serializer
+        self._value_serializer = value_serializer
+
+    @asyncio.coroutine
+    def send(self, topic, value=None, key=None, partition=None,
+             timestamp_ms=None):
+        """Publish a message to a topic.
+
+        Arguments:
+            topic (str): topic where the message will be published
+            value (optional): message value. Must be type bytes, or be
+                serializable to bytes via configured value_serializer. If value
+                is None, key is required and message acts as a 'delete'.
+                See kafka compaction documentation for more details:
+                http://kafka.apache.org/documentation.html#compaction
+                (compaction requires kafka >= 0.8.1)
+            partition (int, optional): optionally specify a partition. If not
+                set, the partition will be selected using the configured
+                'partitioner'.
+            key (optional): a key to associate with the message. Can be used to
+                determine which partition to send the message to. If partition
+                is None (and producer's partitioner config is left as default),
+                then messages with the same key will be delivered to the same
+                partition (but if key is None, partition is chosen randomly).
+                Must be type bytes, or be serializable to bytes via configured
+                key_serializer.
+            timestamp_ms (int, optional): epoch milliseconds (from Jan 1 1970
+                UTC) to use as the message timestamp. Defaults to current time.
+
+        Returns:
+            asyncio.Future: object that will be set when message is
+            processed
+
+        Raises:
+            kafka.KafkaTimeoutError: if we can't schedule this record (
+                pending buffer is full) in up to `request_timeout_ms`
+                milliseconds.
+
+        Note:
+            The returned future will wait based on `request_timeout_ms`
+            setting. Cancelling the returned future **will not** stop event
+            from being sent, but cancelling the ``send`` coroutine itself
+            **will**.
+        """
+        if value is None and self.api_version < (0, 8, 1):
+            raise ValueError('Null messages require kafka >= 0.8.1')
+        if key is value is None:
+            raise ValueError('Need at least one: key or value')
+
+        # first make sure the metadata for the topic is available
+        yield from self.client._wait_on_metadata(topic)
+
+        key_bytes, value_bytes = self._serialize(topic, key, value)
+        partition = self._partition(topic, partition, key, value,
+                                    key_bytes, value_bytes)
+
+        tp = TopicPartition(topic, partition)
+        log.debug("Sending (key=%s value=%s) to %s", key, value, tp)
+
+        fut = yield from self._add_message(
+            tp, key_bytes, value_bytes, self._request_timeout_ms / 1000,
+            timestamp_ms=timestamp_ms)
+        return fut
+
+    @asyncio.coroutine
+    def send_and_wait(self, topic, value=None, key=None, partition=None):
+        """Publish a message to a topic and wait the result"""
+        future = yield from self.send(topic, value, key, partition)
+        return (yield from future)
+
+    def _partition(self, topic, partition, key, value,
+                   serialized_key, serialized_value):
+        if partition is not None:
+            if partition < 0:
+                raise ValueError("Partition may not be negative")
+            if partition not in self._metadata.partitions_for_topic(topic):
+                raise UnknownPartition(partition)
+            return partition
+
+        all_partitions = list(self._metadata.partitions_for_topic(topic))
+        available = list(self._metadata.available_partitions_for_topic(topic))
+        return self._partitioner(serialized_key, all_partitions, available)
+
+    def _serialize(self, topic, key, value):
+        if self._key_serializer:
+            serialized_key = self._key_serializer(key)
+        else:
+            serialized_key = key
+        if self._value_serializer:
+            serialized_value = self._value_serializer(value)
+        else:
+            serialized_value = value
+
+        message_size = MessageSet.HEADER_SIZE + Message.HEADER_SIZE
+        if serialized_key is not None:
+            message_size += len(serialized_key)
+        if serialized_value is not None:
+            message_size += len(serialized_value)
+        if message_size > self._max_request_size:
+            raise MessageSizeTooLargeError(
+                "The message is %d bytes when serialized which is larger than"
+                " the maximum request size you have configured with the"
+                " max_request_size configuration" % message_size)
+
+        return serialized_key, serialized_value
+
+    @asyncio.coroutine
+    def _add_message(self, tp, key, value, timeout, timestamp_ms=None):
+        """ Add message to batch by topic-partition
+        If batch is already full this method waits (`timeout` seconds maximum)
+        until batch is drained by send task
+        """
+        while timeout > 0:
+            if self._closed:
+                raise ProducerClosed()
+
+            batch = self.get_batch(tp)
+            future = batch.append(key, value, timestamp_ms, future=True)
+            if future is not None:
+                return future
+
+            # Batch is full, wait for it to drain
+            start = self._loop.time()
+            yield from batch.wait_drain(timeout=timeout)
+            timeout -= self._loop.time() - start
+
+        raise KafkaTimeoutError()
